Refactor utility functions into separate class
